Poll monthly usage
